move legacy models to a separate namespace, rename
